Extract comments and primitive literals modules
First step of extracting modules to make it easier to navigate the
parsing code.
